Create README.md
a feature-enhanced attention neural network (Fe-Net) to handle the fine-grained image recognition of crop pests & diseases in innovative agronomy practices. This model is estab-lished on the basis of an improved CSP-stage backbone network, which offers massive chan-nel-shuffled features in various dimensions and sizes. Then, a spatial feature-enhanced attention module is added to exploit the spatial interrelationship between different semantic regions. Finally, the proposed Fe-Net employs a higher-order pooling module to mine more high-representative features by computing the square root of the covariance matrix of elements. The whole architec-ture is efficiently trained in an end-to-end way without additional manipulation.